Phase 3: L4-Klangelement, Vorhoer-Renderer, MIDI-Export
Feldrelative Ansteuerung (au/dsl/field.py: Modi, Stufe->MIDI, Transposition)
und Ereignisgeneratoren (Poisson-Dichte, euklidische Verteilung). Ein neues
Modul mod.ctrl.constant bruckt Tonhoehe pro Notenereignis ueber den
bestehenden Makro-Mechanismus in den Compiler, ohne diesen zu aendern.

ElementRecipe (au/dsl/element.py) bindet eine L2-Stimme an Feld+Pattern.
Renderer plant jedes Ereignis als eigene Synth-Instanz mit eigener Tonhoehe
und Attack/Release-Huellkurve im selben NRT-Score (au/render/element.py).

Vorhoeren: RMS-Normalisierung auf einen Zielpegel fuer fairen Vergleich
(au/render/audition.py, spaeter durch echte R128-Lautheit ersetzbar).
MIDI-Export der Ansteuerung (au/modules/io/midi_export.py). Autokorrelations-
basierter Loop-Detektor auf der Envelope (au/analysis/metrics.py).

Bewusst reduzierter Umfang gegenueber plan.md 10.2: eine Symbolikkette/Makro-
Trajektorie pro Element, keine Effektkette, keine Feld-/Kontext-Audition
(brauchen Blueprint bzw. Nachbarelemente aus Phase 6/8). Rezept bleibt
unveraendert bei Transposition (model_copy), keine gerenderte Datei als
Quelle -- Kernentscheidung aus plan.md 10.3 eingehalten.

9 neue Tests (Transposition -7/0/+7, MIDI-Lesbarkeit, Loop-Detektor-
Gegenprobe), 82 insgesamt gruen.

+++ au/render/backend.py
@@ -242,7 +242,7 @@ async def render_score_async(
 
     frames, warnings = _verify_artifact(
         candidate,
-        expected_frames=int(round(duration * c.audio.sample_rate)),
+        expected_frames=round(duration * c.audio.sample_rate),
         expected_rate=c.audio.sample_rate,
         expected_channels=c.audio.channels,
         exit_code=exit_code,
